Split coverageReport.test.js into 2 suits
